+++ src/modules/notifications/ModuleNotifications.vue
@@ -87,6 +87,7 @@ import { formatFloatingPointValue } from '@/utils/numberFormatHelper'
 import { ToastType } from '@/types/notification'
 import Configs from '@/configs'
 import { captureException } from '@sentry/vue'
+import { SENTRY_MODULE_TAGS } from '@/sentry/constants'
 
 // Stores
 import { storeToRefs } from 'pinia'
@@ -437,7 +438,7 @@ const startPolling = async (hash: string, chainId: number) => {
       const status = await fusion.getOrderStatus(hash)
       updateOrderStatus(hash, status)
     } catch (e) {
-      captureException(e)
+      captureException(e, SENTRY_MODULE_TAGS.NOTIFICATIONS)
     }
   }
 
@@ -691,7 +692,7 @@ const startStatusPolling = async (
         updateNotificationStatus(hash, type, data)
       }
     } catch (e) {
-      captureException(e)
+      captureException(e, SENTRY_MODULE_TAGS.NOTIFICATIONS)
     }
   }